Config Server (#5)
This PR introduces a number of items as well as a few fixes. The primary purpose of the PR is to add a configuration parameter microservice that can be queried and manipulated via http and json. This allows for agnostic access on the network. Convenience python wrappers are provided in `panoptes_utils/config/client.py` that automatically handled serialization into preferred objects.

* Change default database type
  * **Behavior Change** This changes the default database type to `memory` so that nothing is
required to use this inside a console. Note that POCS will still set
a default of `file`.