Initial commits for some of the data access components.
`get_data` is a wrapper around the more specific functions. For now that is `get_image` and `get_observation`.

`get_image` is working with this commit. Currently it does a lookup on the firestore db in order to get the full path (because the `sequence_id` is not part of the `image_id` and then uses the google client to download files.

Alternatively we could create a cloud function that hides the details of firestore (so the user doesn't need login access). Likewise, we could download via the public url and `requests` rather than use the google storage libraries.